Retry VT_WAITACTIVE ioctl if interrupted
My personal use-case that triggers the problem:
physlock is launched by the XFCE Power Manager daemon when the system is
suspended. What happens is that, sometimes, the WAITACTIVE ioctl is
issued just before the system goes to sleep. When the system resumes,
the ioctl returns with errno=EINTR (aka interrupted system call),
therefore physlock exits.

Retrying the ioctl is enough to solve the problem.