Add option to mute kernel messages on console
My personal use-case:
I lock my computer with physlock automatically when the computer is
suspended. On resume, the kernel spits plenty of log messages to the
console, and the physlock prompt is lost inside this mess.

With an option to mute kernel logs messages, we can ensure that there's
nothing but the physlock prompt on the console, and it looks a lot
friendlier.